Coil-Over Model Files
Used Fusion 360 to model a self-designed coil-over. Iterated to find a method which worked. Initially, there was a screw design for the piston to be inserted into the base, but pins replaced the screw because the screw was impossible to access. Additionally, piston geometry was fine-tuned and a knob was added on top to add weight to simulate different effects.