Demonstration of Dampened Oscillation: Coil-Over
Using different fluids, we explored the effects of viscosity of fluids on the dampening of harmonic oscillation. We used a baseline test where there was no fluid or any friction whatsoever by using just the spring. From there, we then put the springs in the coil-over assembly and tested how the springs worked with just the friction from the assembly (air as our fluid) and then water. We saw how the water and the assembly quickly dissipated the energy of the spring, whereas the air and the assembly itself took much longer to dissipate the energy. We also noticed how the period of the oscillation was unaffected by the dampened motion, and the amplitudes form or follow an inverse exponential function (e^-x).